1. Modify the structure so that it can be installed simply through "git clone" in the "custom_nodes" path.  - A structure manageable by ComfyUI-Manager.
2. Enhance it to manage the JavaScript path using the WEB_DIRECTORY property supported by the latest version of ComfyUI.
