+++ color_clip_ade20k.py
