+++ cv_nodes.py
