+++ js/ArbitraryInputsFixedOutputs.js
