+++ otsu_threshold_node.py
